Cartan.jl Library

Cartan.Coordinate

Cartan.Coordinates

Cartan.DiscontinuousTopology

Cartan.FaceBundle

Cartan.FiberBundle

Cartan.FiberProduct

Cartan.FiberProductBundle

Cartan.FrameBundle

Cartan.Global

Cartan.GlobalFiber

Cartan.GridBundle

Cartan.HomotopyBundle

Cartan.ImmersedTopology

Cartan.LocalFiber

Cartan.LocalTensor

Cartan.PointArray

Cartan.ProductSpace

Cartan.ProductTopology

Cartan.QuotientTopology

Cartan.SimplexBundle

Cartan.SimplexTopology

Cartan.TensorField

Cartan.coordinates

Cartan.immersion

Cartan.base

Cartan.basetype

Cartan.bundle

Cartan.continuous

Cartan.coordinatetype

Cartan.disconnect

Cartan.discontinuous

Cartan.elements

Cartan.fiber

Cartan.fibertype

Cartan.fullcoordinates

Cartan.fullimmersion

Cartan.fullmetricextensor

Cartan.fullpoints

Cartan.fulltopology

Cartan.fullvertices

Cartan.getfacet

Cartan.getimage

Cartan.immersiontype

Cartan.iscover

Cartan.isdisconnected

Cartan.isdiscontinuous

Cartan.isfull

Cartan.isinduced

Cartan.istotal

Cartan.metrictype

Cartan.nodes

Cartan.pointtype

Cartan.refnodes

Cartan.resample

Cartan.sdims

Cartan.subelements

Cartan.subimmersion

Cartan.topology

Cartan.totalelements

Cartan.totalnodes

Cartan.vertices

DirectSum.:⊕

Grassmann.points

Cartan.Coordinate — Type

Coordinate{P,G} <: LocalFiber{P,G} <: Number

Defines a Coordinate bundled type with pointtype of P and metrictype of G.

point(s) # ::P
metricextensor(s) # ::G
pointtype(s) # P
metrictype(s) # G

A Coordinate{P,G} consists of two components: P, which represents the point manifold, and G, which represents the metricextensor bundle over P.

Cartan.Coordinates — Type

Coordinates{P,G,N} <: FiberBundle{Coordinate{P,G},N} <: Number

Defines a FiberBundle type with pointtype of P and metrictype of G.

coordinates(s) # ::AbstractArray{Coordinate{P,G},N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # Coordinate{P,G}
pointtype(s) # P
metrictype(s) # G

Various methods work on any Coordinates, such as base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype.

Cartan.DiscontinuousTopology — Type

DiscontinuousTopology{N} <: ImmersedTopology{N,1}

Defines discontinuous subspace of a fulltopology over N-dimensional Simplex spaces having vertices and subelements defined and indexed in its simplicial complex.

isdiscontinuous(t) # true if DiscontinuousTopology
isdisconnected(t) # true if graph of discontinuous

Calling DiscontinuousToppology(::SimplexTopology) initializes a discontinuous variant, while SimplexTopology(::DiscontinuousTopology) returns the initial continuous data. Related methods include bundle, fulltopology, topology, fullvertices, vertices, totalelements, elements, subelements, totalnodes, nodes, istotal, isfull, iscover, fullimmersion, subimmersion, isdiscontinuous, isdisconnected, continuous, discontinuous, disconnect.

Cartan.FaceBundle — Type

FaceBundle{N,C} <: FrameBundle{C,1} <: FiberBundle{C,1}

Defines a FrameBundle over element faces with coordinatetype of C and immersion.

coordinates(s) # ::AbstractArray{C,N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # C
pointtype(s) # P
metrictype(s) # G
immersion(s) # ::ImmersedTopology

Various methods work on any FrameBundle, such as isbundle,base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype, immersion.

Cartan.FiberBundle — Type

FiberBundle{T,N} <: Number

Defines a global FiberBundle type with basetype and fibertype over N dimensions.

Cartan.FiberProduct — Type

FiberProduct{P,N} <: FiberBundle{Coordinate{P,InducedMetric},N}

Represents a FiberProduct over a base and fiber with InducedMetric.

Cartan.FiberProductBundle — Type

FiberProductBundle{P,N} <: FrameBundle{Coordinate{P,InducedMetric},N}

Represents a FiberProductBundle over a base and fiber with InducedMetric.

Cartan.FrameBundle — Type

FrameBundle{C,N} <: GlobalFiber{C,N} <: FiberBundle{C,N}

Defines a GlobalFiber type with coordinatetype of C and immersion.

coordinates(s) # ::AbstractArray{C,N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # C
pointtype(s) # P
metrictype(s) # G
immersion(s) # ::ImmersedTopology

Various methods work on any FrameBundle, such as isbundle, base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype, immersion.

Cartan.Global — Type

Global{N,T} <: AbstractArray{T,N}

Represents an AbstractArray where every local value is globally the same.

julia> Global{1}(InducedMetric())
Global{1}(InducedMetric())

julia> ans[1]
InducedMetric()

For example, Global{N,InducedMetric} is commonly used for a globally induced metric.

Cartan.GlobalFiber — Type

GlobalFiber{E,N} <: FiberBundle{E,N} <: Number

Defines a FiberBundle type with basetype and fibertype.

Cartan.GridBundle — Type

GridBundle{N,C} <: FrameBundle{C,N} <: FiberBundle{C,N}

Defines a FrameBundle over grid points with coordinatetype of C and immersion.

coordinates(s) # ::AbstractArray{C,N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # C
pointtype(s) # P
metrictype(s) # G
immersion(s) # ::QuotientTopology

Various methods work on any FrameBundle, such as isbundle, base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype, immersion.

IntervalRange{P, G, PA, GA} where {P<:Real, G, PA<:AbstractRange, GA} (alias for GridBundle{1, Coordinate{P, G}, <:PointArray{P, G, 1, PA, GA}} where {P<:Real, G, PA<:AbstractRange, GA})
AlignedRegion{N} where N (alias for GridBundle{N, Coordinate{P, G}, PointArray{P, G, N, PA, GA}} where {N, P<:Chain, G<:InducedMetric, PA<:(ProductSpace{V, <:Real, N, N, <:AbstractRange} where V), GA<:Global})
AlignedSpace{N} where N (alias for GridBundle{N, Coordinate{P, G}, PointArray{P, G, N, PA, GA}} where {N, P<:Chain, G<:InducedMetric, PA<:(ProductSpace{V, <:Real, N, N, <:AbstractRange} where V), GA})

Cartan.HomotopyBundle — Type

HomotopyBundle{P,N} <: FrameBundle{Coordinate{P,InducedMetric},N}

Represents a HomotopyBundle over a base and fiber with InducedMetric.

Cartan.ImmersedTopology — Type

ImmersedTopology{N,M} = AbstractArray{Values{N,Int},M}

Any ImmersedTopology{N,M} is also an M-dimensional AbstractArray of Values{N,Int}.

Cartan.LocalFiber — Type

LocalFiber{B,F} <: Number

Defines abstract bundled type with basetype of B and fibertype of F in a manifold.

base(s) # ::B
fiber(s) # ::F
basetype(s) # B
fibertype(s) # F

A LocalFiber{B,F} consists of two components: B, which represents the base manifold, and F, which represents the fiber bundle over B.

Cartan.LocalTensor — Type

LocalTensor{B,F} <: LocalFiber{B,F} <: Number

Defines a local bundled type with basetype of B and fibertype of F.

base(s) # ::B
fiber(s) # ::F
basetype(s) # B
fibertype(s) # F

A LocalTensor{B,F} consists of two components: B, which represents the base manifold, and F, which represents the fiber bundle over B.

Cartan.PointArray — Type

PointArray{P,G,N} <: FiberBundle{Coordinate{P,G},N} <: Number

Defines a FiberBundle type with pointtype of P and metrictype of G.

coordinates(s) # ::PointArray{P,G,N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # ::Coordinate{P,G}
pointtype(s) # P
metrictype(s) # G

Various methods work on any PointArray, such as base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype.

Cartan.ProductSpace — Type

ProductSpace{V,T,N,M,S} <: AbstractArray{Chain{V,1,T,N},N}

Can be constructed with \oplus operation ⊕ and AbstractRange,

julia> (0:0.1:1)⊕(0:0.1:1)
11×11 ProductSpace{⟨_11_⟩, Float64, 2, 2, StepRangeLen{Float64, Base.TwicePrecision{Float64}, Base.TwicePrecision{Float64}, Int64}}:
...

generating a lazy array of Chain{V,1} point vectors from the input ranges.

Rectangle (alias for ProductSpace{V, T, 2, 2} where {V, T})
Hyperrectangle (alias for ProductSpace{V, T, 3, 3} where {V, T})
RealRegion{V, T} where {V, T<:Real} (alias for ProductSpace{V, T, N, N, S} where {V, T<:Real, N, S<:AbstractArray{T, 1}})

Cartan.ProductTopology — Type

ProductTopology{N} <: ImmersedTopology{N,N}

Define basic ProductTopology by ranges of integers,

julia> ProductTopology(11,11)
11×11 ProductTopology{2, OneTo{Int64}}:
...

julia> ProductTopology(1:11,1:11)
11×11 ProductTopology{2, UnitRange{Int64}}:
...

Cartan.QuotientTopology — Type

QuotientTopology{N} <: ImmersedTopology{N,N}

Generalizing upon the ProductTopology, the QuotientTopology defines a quotient identification across the boundary fluxes of the region, from which the differential topology induced compact local substructure is derived.

isopen(t) # true if open topology
iscompact(t) # true if compact topology

Common instances include OpenTopology, CompactTopology, RibbonTopology, MobiusTopology, WingTopology, MirrorTopology, ClampedTopology, TorusTopology, HopfTopology, KleinTopology, ConeTopology, PolarTopology, SphereTopology, GeographicTopology.

Cartan.SimplexBundle — Type

SimplexBundle{N,C} <: FrameBundle{C,1} <: FiberBundle{C,1}

Defines a FrameBundle over simplex vertices with coordinatetype of C and immersion.

coordinates(s) # ::AbstractArray{C,N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # C
pointtype(s) # P
metrictype(s) # G
immersion(s) # ::ImmersedTopology

Various methods work on any FrameBundle, such as isbundle, base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype, immersion.

Cartan.SimplexTopology — Type

SimplexTopology{N} <: ImmersedTopology{N,1}

Defines continuous subspaces of a fulltopology over N-dimensional Simplex spaces having vertices and subelements defined and indexed in its simplicial complex.

bundle(t) # cache identification
fulltopology(t) # full element list
vertices(t) # # subspace vertices
totalnodes(t) # full node count
subelements(t) # list of elements
fullvertices(t) # fulltopology vertices
verticesinv(t) # inverted data of vertices

Related methods include bundle, fulltopology, topology, fullvertices, vertices, totalelements, elements, subelements, totalnodes, nodes, istotal, isfull, iscover, fullimmersion, subimmersion.

Cartan.TensorField — Type

TensorField{B,F,N} <: GlobalFiber{LocalTensor{B,F},N}

Defines a GlobalFiber type with eltype of LocalTensor{B,F} and immersion.

coordinates(s) # ::AbstractArray{B,N}
fiber(s) # ::AbstractArray{F,N}
points(s) # ::AbstractArray{P,N}
metricextensor(s) # ::AbstractArray{G,N}
coordinatetype(s) # B
fibertype(s) # F
pointtype(s) # P
metrictype(s) # G
immersion(s) # ::ImmersedTopology

Various methods work on any TensorField, such as base, fiber, coordinates, points, metricextensor, basetype, fibertype, coordinatetype, pointtype, metrictype, immersion. Due to the versatility of the TensorField type instances, it's possible to disambiguate them into these type alias specifications with associated methods: ElementMap, SimplexMap, FaceMap, IntervalMap, RectangleMap, HyperrectangleMap, ParametricMap, Variation, RealFunction, PlaneCurve, SpaceCurve, AbstractCurve, SurfaceGrid, VolumeGrid, ScalarGrid, DiagonalField, EndomorphismField, OutermorphismField, CliffordField, QuaternionField, ComplexMap, PhasorField, SpinorField, GradedField{G} where G, ScalarField, VectorField, BivectorField, TrivectorField.

In the Cartan package, a technique is employed where a TensorField is constructed from an interval, product manifold, or topology, to generate an algebra of sections which can be used to compose parametric maps on manifolds. Constructing a TensorField can be accomplished in various ways, there are explicit techniques to construct a TensorField as well as implicit methods. Additional packages such as Adapode build on the TensorField concept by generating them from differential equations. Many of these methods can automatically generalize to higher dimensional manifolds and are compatible with discrete differential geometry.

Cartan.coordinates — Type

coordinates(m::FiberBundle) -> FiberBundle{Coordinate{P,G}}

Return a FiberBundle{Coordinate{P,G}} if object m is defined as a FiberBundle.

Cartan.immersion — Type

immersion(m) -> ImmersedTopology

Returns the associated ImmersedTopology of any FrameBundle or related object.

Cartan.base — Method

base(m::LocalFiber{B,F}) -> B

Return the base of a FiberBundle or LocalFiber.

Cartan.basetype — Method

basetype(m::LocalFiber{B,F}) -> DataType

Return the basetype of a FiberBundle or LocalFiber.

Cartan.bundle — Method

bundle(m::ImmersedTopology) -> Int

Returns integer identification of bundle cache.

Cartan.continuous — Method

continuous(m) -> ImmersedTopology

Return the original continuous data of possibly a DiscontinuousTopology.

Cartan.coordinatetype — Method

coordinatetype(m::Coordinate{P,G}) -> DataType

Return the coordinatetype of a FiberBundle or LocalFiber.

Cartan.disconnect — Method

disconnect(m) -> DiscontinuousTopology

Return a disconnected variant of a DiscontinuousTopology.

Cartan.discontinuous — Method

discontinuous(m) -> DiscontinuousTopology

Return a derived DiscontinuousTopology from a SimplexTopology.

Cartan.elements — Method

elements(m) -> Int

Return the number of subelements(m).

Cartan.fiber — Method

fiber(m::LocalFiber{B,F}) -> F

Return the fiber of a FiberBundle or LocalFiber.

Cartan.fibertype — Method

fibertype(m::LocalFiber{B,F}) -> DataType

Return the fibertype of a FiberBundle or LocalFiber.

Cartan.fullcoordinates — Method

fullcoordinates(m::FiberBundle) -> FiberBundle{Coordinate{P,G}}

Return full FiberBundle{Coordinate{P,G}} instead of a possible subspace of it.

Cartan.fullimmersion — Method

fullimmersion(m) -> ImmersedTopology

Returns full associated ImmersedTopology of subspace FrameBundle or related object.

Cartan.fullmetricextensor — Method

fullmetricextensor(m::FiberBundle) -> AbstractArray{G}

Return full AbstractArray{G} instead of a possible subspace of it.

Cartan.fullpoints — Method

fullpoints(m::FiberBundle) -> AbstractArray{P}

Return full AbstractArray{P} instead of a possible subspace of it.

Cartan.fulltopology — Method

fulltopology(m) -> Vector{Values{N,Int}}

Returns the fulltopology regardless of whether the ImmersedTopology subset isfull.

Cartan.fullvertices — Method

fullvertices(m) -> AbstractVector{Int}

Return the list of vertices(m) associated to the fullimmersion(m).

Cartan.getfacet — Method

getfacet(m,i) -> Int

Return the index of subelements in reference to the fullimmersion.

Cartan.getimage — Method

getimage(m,i) -> Int

Return the index of vertex subspace immersion in reference to the fullimmersion.

Cartan.immersiontype — Method

immersiontype(m) -> DataType

Returns the typeof the immersion of m.

Cartan.iscover — Method

iscover(m::SimplexTopology) -> Bool

Return true if isfull(m) && istotal(m), and false otherwise.

Cartan.isdisconnected — Method

isdiscontinuous(m) -> Bool

Return true if m is a disconnected DiscontinuousTopology.

Cartan.isdiscontinuous — Method

isdiscontinuous(m) -> Bool

Return true if m is a DiscontinuousTopology.

Cartan.isfull — Method

isfull(m::SimplexTopology) -> Bool

Return true if m is equal to fulltopology(m), and false otherwise.

Cartan.isinduced — Method

isinduced(m) -> Bool

Return true if the metrictype is an InducedMetric type.

Cartan.istotal — Method

istotal(m::SimplexTopology) -> Bool

Return true if fulltopology(m) is covering totalnodes(m), and false otherwise.

Cartan.metrictype — Method

metrictype(m::Coordinate{P,G}) -> DataType

Return the metrictype of a FiberBundle or LocalFiber.

Cartan.nodes — Method

nodes(m) -> Int

Return the number of vertices(m) associated to the subspace immersion(m).

Cartan.pointtype — Method

pointtype(m::Coordinate{P,G}) -> DataType

Return the pointtype of a FiberBundle or LocalFiber.

Cartan.refnodes — Method

refnodes(m) -> Base.RefValaue{Int}

Return the shared mutable Base.RefValue{Int} which counts nodes(m).

Cartan.resample — Method

resample(m,i...)

Resamples a ranged ProductSpace or related objects.

Cartan.sdims — Method

sdims(::ImmersedTopology{N}) where N = N

Dimension N of the associated immersed simplex.

Cartan.subelements — Method

subelements(m) -> AbstractVector{Int}

Return the subspace element indices associated to fulltopology(m).

Cartan.subimmersion — Method

subimmersion(m) -> ImmersedTopology

Return a modified subimmersion with all vertices re-indexed based on the subspace.

Cartan.topology — Method

topology(m) -> AbstractVector{Values{N,Int}}

Returns a view into the fulltopology based on subelements(m) structure.

Cartan.totalelements — Method

totalelements(m) -> Int

Return the total number of elements in the fulltopology(m).

Cartan.totalnodes — Method

totalnodes(m) -> Int

Return the number which counts the total number of nodes regardless of subspace.

Cartan.vertices — Method

vertices(m) -> AbstractVector{Int}

Return the list of vertices(m) associated to the subspace immersion(m).

DirectSum.:⊕ — Method

⊕(v::AbstractVector{<:Real}...)

Constructs a direct sum basis space using the Cartesian ProductSpace implementation.

Grassmann.points — Method

points(m) -> AbstractArray{P}

Return the points as an AbstractArray{P}.